core/requirements.txt
Django>=5.2.5
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-environ>=0.11.2
python-dotenv>=1.0.0
PyJWT>=2.8.0
cryptography>=41.0.0
psycopg2-binary>=2.9.7
channels>=4.0.0
channels-redis>=4.1.0
djangorestframework-simplejwt>=5.3.0
whitenoise>=6.5.0
pillow>=10.0.0
stripe>=5.0.0
paypalhttp>=1.0.1
drf-yasg>=1.21.7